Dev (#7)
* refactor(workflows): update build.yml to include workflow_dispatch

Added a workflow_dispatch trigger to the build process, enabling manual execution of the workflow. Also, adjusted the formatting of the table in the GitHub step summary to ensure better alignment.

Signed-off-by: Liz C. <alicecfire@gmail.com>

* chore(workflows): correct version update command in build workflow (dev)

Updated the version update command in the build workflow to ensure it properly references the output from the versioning job. This modification corrects the way the version is set in the gradle.properties file during the CI process.

Signed-off-by: Liz C. <alicecfire@gmail.com>

* chore(gitVersion): update version bump messages

Updated the version bump messages in the gitVersion.yml file to use multi-line format for better readability. The major version bump message remains unchanged, while the minor and patch version bump messages have been revised to allow for multi-line definitions.

Signed-off-by: Liz C. <alicecfire@gmail.com>

* chore(workflow): correct echo command syntax in build.yml

Updated the echo command in the GitHub Actions workflow to ensure proper syntax. The change modifies the line to use double quotes for better handling of the output from the gitversion step.

Signed-off-by: Liz C. <alicecfire@gmail.com>

* chore(workflow): update echo command syntax in build.yml (dev)

Updated the echo command in the print outputs step to use single quotes for consistency and to avoid potential issues with string interpolation.
This change ensures that the output is printed correctly without any
unintended formatting.

Issue: #dev
Signed-off-by: Liz C. <alicecfire@gmail.com>

* chore(workflow): correct syntax in build.yml

Fixed a syntax error in the build.yml file where the version
updating command was missing a closing quote. This change ensures
that the version is correctly updated in the gradle.properties file.

Signed-off-by: Liz C. <alicecfire@gmail.com>

* chore(workflow): add pull request trigger and update permissions (dev)

- Added a pull request trigger to the build workflow for the src path.
- Updated permissions to allow write access for the contents.

Signed-off-by: Liz C. <alicecfire@gmail.com>

* chore(gradle): update version to 1.0.0-alpha.49

Updated the version number from 1.0.1  to 1.0.0-alpha.49 in the gradle.properties file to reflect  the latest release.

---------

Signed-off-by: Liz C. <alicecfire@gmail.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>