build(nyx): move "INTERNAL" to the end of the "enabled" release stages list (dev)
This commit updates the .nyx.json configuration to place "INTERNAL" at the end of the list of enabled release stages. This change improves the management of the release process and ensures that internal releases can be properly tracked and utilized. The addition of "INTERNAL-DIRTY" also allows for better handling of builds in a dirty workspace state.

Signed-off-by: Liz C. <alicecfire@gmail.com>